Cleanup: Comment & Subreddit request.
