+++ jraw/pom.xml
@@ -22,7 +22,7 @@
     <parent>
         <artifactId>reddit</artifactId>
         <groupId>vartas.reddit</groupId>
-        <version>3.1</version>
+        <version>3.2-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
