+++ jraw/src/main/java/vartas/reddit/JrawSubmission.java
@@ -17,11 +17,17 @@
 
 package vartas.reddit;
 
+import net.dean.jraw.RedditClient;
+import net.dean.jraw.references.SubmissionReference;
+import net.dean.jraw.tree.RootCommentNode;
 import org.apache.commons.text.StringEscapeUtils;
 import org.apache.commons.validator.routines.UrlValidator;
+import org.slf4j.LoggerFactory;
 import vartas.reddit.factory.SubmissionFactory;
 
 import javax.annotation.Nonnull;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 @Nonnull
@@ -36,7 +42,7 @@ public class JrawSubmission extends Submission {
     }
 
     @Nonnull
-    public static Submission create(@Nonnull net.dean.jraw.models.Submission jrawSubmission){
+    public static Submission create(@Nonnull net.dean.jraw.models.Submission jrawSubmission, RedditClient jrawClient){
         Submission submission = SubmissionFactory.create(
                 () -> new JrawSubmission(jrawSubmission),
                 jrawSubmission.getAuthor(),
@@ -54,8 +60,33 @@ public class JrawSubmission extends Submission {
 
         submission.setLinkFlairText(Optional.ofNullable(jrawSubmission.getLinkFlairText()));
         submission.setContent(Optional.ofNullable(jrawSubmission.getSelfText()).map(StringEscapeUtils::unescapeHtml4));
+        submission.addAllComments(requestComments(jrawSubmission, jrawClient));
 
         return submission;
+    }
+
+    private static List<Comment> requestComments(net.dean.jraw.models.Submission jrawSubmission, RedditClient jrawClient){
+        List<Comment> comments = new ArrayList<>();
+
+        RootCommentNode root;
+        SubmissionReference submissionReference = jrawSubmission.toReference(jrawClient);
+
+        try {
+            root = submissionReference.comments();
+            //Acquire all the comments
+            root.loadFully(jrawClient);
+
+            root.walkTree().iterator().forEachRemaining(node -> {
+                if(node.getSubject() instanceof net.dean.jraw.models.Comment)
+                    comments.add(JrawComment.create((net.dean.jraw.models.Comment)node.getSubject()));
+            });
+
+        }catch(NullPointerException e){
+            //null if the submission doesn't exist -> Not a communication error
+            LoggerFactory.getLogger(JrawSubmission.class.getSimpleName()).warn(e.getMessage(), e);
+        }
+
+        return comments;
     }
 
     @Nonnull
